fastapi-backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
python-dotenv==1.0.0
httpx==0.24.1
python-multipart==0.0.6
uuid==1.30